+++ tests/integration/mask.rs
@@ -207,3 +207,57 @@ fn ignore_source() {
     let expected = Pixmap::load_png("tests/images/mask/ignore-source.png").unwrap();
     assert_eq!(pixmap, expected);
 }
+
+#[test]
+fn apply_mask() {
+    let mut pixmap = Pixmap::new(100, 100).unwrap();
+
+    let clip_path = PathBuilder::from_circle(100.0, 100.0, 50.0).unwrap();
+    let mut mask = Mask::new(100, 100).unwrap();
+    mask.fill_path(&clip_path, FillRule::Winding, true, Transform::default());
+
+    let mut paint = Paint::default();
+    paint.set_color_rgba8(50, 127, 150, 200);
+    paint.anti_alias = false;
+
+    let rect = Rect::from_xywh(0.0, 0.0, 100.0, 100.0).unwrap();
+    pixmap.fill_rect(rect, &paint, Transform::identity(), None);
+    pixmap.apply_mask(&mask);
+
+    let expected = Pixmap::load_png("tests/images/mask/apply-mask.png").unwrap();
+    assert_eq!(pixmap, expected);
+}
+
+#[test]
+fn mask_from_alpha() {
+    let path = PathBuilder::from_circle(100.0, 100.0, 50.0).unwrap();
+
+    let mut paint = Paint::default();
+    paint.set_color_rgba8(50, 127, 150, 200);
+    paint.anti_alias = true;
+
+    let mut pixmap = Pixmap::new(100, 100).unwrap();
+    pixmap.fill_path(&path, &paint, FillRule::Winding, Transform::default(), None);
+
+    let mask = Mask::from_pixmap(pixmap.as_ref(), MaskType::Alpha);
+
+    let expected = Mask::load_png("tests/images/mask/mask-from-alpha.png").unwrap();
+    assert_eq!(mask, expected);
+}
+
+#[test]
+fn mask_from_luma() {
+    let path = PathBuilder::from_circle(100.0, 100.0, 50.0).unwrap();
+
+    let mut paint = Paint::default();
+    paint.set_color_rgba8(50, 127, 150, 200);
+    paint.anti_alias = true;
+
+    let mut pixmap = Pixmap::new(100, 100).unwrap();
+    pixmap.fill_path(&path, &paint, FillRule::Winding, Transform::default(), None);
+
+    let mask = Mask::from_pixmap(pixmap.as_ref(), MaskType::Luminosity);
+
+    let expected = Mask::load_png("tests/images/mask/mask-from-luma.png").unwrap();
+    assert_eq!(mask, expected);
+}